restructure project move main.go to /example make this project to be a slip package
